aotf: restructure and test aotf logic
* separate pure graphql logic from aotf
