+++ src/utils/graphql.js
@@ -15,7 +15,6 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import dedent from 'dedent'
 import {
   ApolloClient,
   ApolloLink,
@@ -27,7 +26,6 @@ import { getMainDefinition } from '@apollo/client/utilities'
 import { WebSocketLink } from '@apollo/client/link/ws'
 import { SubscriptionClient } from 'subscriptions-transport-ws'
 import store from '@/store'
-import gql from 'graphql-tag'
 
 /**
  * Create the HTTP and WebSocket URLs for an ApolloClient.
@@ -141,153 +139,4 @@ export function createApolloClient (httpUrl, subscriptionClient) {
     },
     connectToDevTools: process.env.NODE_ENV !== 'production'
   })
-}
-
-/** Walk a GraphQL type yielding all composite types on route.
- *
- * E.G. NonNull<List<String>> would yield:
- *  1. NonNull
- *  2. List
- *  3. String
- *
- * @param type {Object} A type as returned by an introspection query.
- * (i.e. an object of the form {name: x, kind: y, ofType: z}
- *
- * @yields {Object} Type objects of the same form as the type argument.
- */
-export function * iterateType (type) {
-  while (type) {
-    yield type
-    type = type.ofType
-  }
-}
-
-/** Return an appropriate null value for the specified type.
- *
- * @param type {Object} A type field as returned by an introspection query.
- * (an object of the form {name: x, kind: y, ofType: z}).
- * @param types {Array} An array of all types present in the schema.
- * (optional: used to resolve InputObjectType fields).
- *
- * @returns {Object|Array|undefined}
- */
-export function getNullValue (type, types = []) {
-  let ret = null
-  let ofType = null
-  for (const subType of iterateType(type)) {
-    if (subType.kind === 'LIST') {
-      ofType = getNullValue(subType.ofType)
-      if (ofType) {
-        // this list contains an object
-        ret = [
-          getNullValue(subType.ofType)
-        ]
-      } else {
-        ret = []
-      }
-      break
-    }
-    if (subType.kind === 'INPUT_OBJECT') {
-      ret = {}
-      for (const type of types) {
-        // TODO: this type iteration is already done in the mixin
-        //       should we use the mixin or a subset there-of here?
-        if (
-          type.name === subType.name &&
-          type.kind === subType.kind
-        ) {
-          for (const field of type.inputFields) {
-            ret[field.name] = getNullValue(field.type)
-          }
-          break
-        }
-      }
-      break
-    }
-  }
-  return ret
-}
-
-/** Return the signature for an argument.
- *
- * E.G: NonNull<List<String>>  =>  [String]!
- *
- * @param arg {Object} An argument from a introspection query.
- *
- * @returns {string} A type string for use in a client query / mutation.
- */
-export function argumentSignature (arg) {
-  const stack = [...iterateType(arg.type)]
-  stack.reverse()
-  let ret = ''
-  for (const type of stack) {
-    if (
-      type.name === null &&
-      type.kind === 'LIST'
-    ) {
-      ret = `[${ret}]`
-    } else if (
-      type.name === null &&
-      type.kind === 'NON_NULL'
-    ) {
-      ret = ret + '!'
-    } else if (type.name) {
-      ret = type.name
-    } else {
-      ret = type.kind
-    }
-  }
-  return ret
-}
-
-/** Construct a mutation string from a mutation introspection.
- *
- * @param mutation {Object} A mutation as returned by an introspection query.
- *
- * @returns {string} A mutation string for a client to send to the server.
- */
-export function constructMutation (mutation) {
-  const argNames = []
-  const argTypes = []
-  for (const arg of mutation.args) {
-    argNames.push(`${arg.name}: $${arg.name}`)
-    argTypes.push(`$${arg.name}: ${argumentSignature(arg)}`)
-  }
-
-  return dedent`
-    mutation ${mutation.name}(${argTypes.join(', ')}) {
-      ${mutation.name}(${argNames.join(', ')}) {
-        result
-      }
-    }
-  `.trim()
-}
-
-// enumeration for the mutation status, maps onto Cylc Task status
-export const status = {
-  waiting: 'waiting',
-  submitted: 'submitted',
-  succeeded: 'succeeded',
-  failed: 'failed',
-  submitFailed: 'submit-failed'
-}
-Object.freeze(status)
-
-export async function mutate (mutation, args, apolloClient) {
-  let result = null
-  try {
-    result = await apolloClient.mutate({
-      mutation: gql(constructMutation(mutation)),
-      variables: args
-    })
-  } catch (err) {
-    // eslint-disable-next-line no-console
-    console.error(err)
-    return [status.submitFailed, err]
-  }
-  const responses = result.data[mutation.name].result
-  if (responses && responses.length === 1) {
-    return [status.submitted, responses[0].response]
-  }
-  return [status.submitFailed, result]
 }
